more tweaks

+++ src/hooks/GJBaseGameLayer/RandomNoJump.cpp
@@ -10,7 +10,7 @@ using namespace horrible;
 class $modify(NoJumpGJBaseGameLayer, GJBaseGameLayer) {
     struct Fields {
         bool enabled = options::get("no_jump");
-        float chance = options::get("no_jump");
+        int chance = options::get("no_jump");
     };
 
     void handleButton(bool down, int button, bool isPlayer1) {
